Update crosstool-ng config files.
Do not use compiled make 3.81, as the build with it often fails (known
error). This could probably cause build failures on new makes, but we
will deal with it when we get there.

Also, we can now use parallel build, which was off because of the build
failures were more often with it.